Task 3

+++ 0x10-variadic_functions/3-print_all.c
@@ -24,8 +24,7 @@ void print_all(const char * const format, ...)
 			{
 				printf(", ");
 				break;
-			}
-			j++;
+			} j++;
 		}
 		switch (format[i])
 		{
@@ -36,7 +35,7 @@ void print_all(const char * const format, ...)
 				printf("%d", va_arg(valist, int)), c = 1;
 				break;
 			case 'f':
-				printf("%f", va_arg(valist, int)), c = 1;
+				printf("%f", va_arg(valist, double)), c = 1;
 				break;
 			case 's':
 				str = va_arg(valist, char *), c = 1;
@@ -47,7 +46,7 @@ void print_all(const char * const format, ...)
 				}
 				printf("%s", str);
 				break;
-		}	i++;
+		} i++;
 	}
 	printf("\n"), va_end(valist);
 }
